Support fetching channel info on trigger
This is acheived with two new methods: TriggerWithParams and
TriggerMultiWithParams which allow an arbitrary map of params to
be passed in. This is how the `info` parameter will be specified.

These methods should be used in favour of TriggerExclusive and
TriggerMultiExclusive - the "socket_id" can be passed in as one
of the parameters in TriggerWithParams instead.

The Trigger*WithParams methods return a `TriggerChannelsList`
which is populated with channel attributes if the `info` parameter
is specified. I considered re-using the existing ChannelsList
type, but I was concerned that the set of required/optional
fields is not quite the same, and may diverge in the future.

Another Approach could have been to drop these return types
entirely and return a weakly-typed map of attribute to value.
I think this might be nicer, but it would imply a larger refactor
and a breaking change to the existing "GET" methods.