Add support fetching channel info to TriggerBatch
This is a breaking change because it changes the return type of
this method to include the returned channel attributes.

This could have been avoided by introducing a new method
`TriggerBatchResponse`, but I didn't think it was worth the
additional complexity in the API.

Another potentially controversial decision was to re-use the same
return type as Trigger*WithParams. This is ok for now when the
responses are identical, but in the future its feasible that
Channels might return different different responses for each of
these endpoints. The library would require a breaking change if
this was ever done. I think thats worth risking in order to keep
things simpler for now.