v5.0.0 release
Note that while we already have a go.mod and go.sum files, this
package was not fully compatible with go modules because the
pacakge name had not been updated to
github.com/pusher/pusher-http-go/v4. This commit completes this
step.

We are taking the approach described in option 1. of
https://github.com/golang/go/wiki/Modules#releasing-modules-v2-or-higher,
and leaving major versions on the master branch.

Note that this will cause problems for legacy package managers
like `dep`.